mattermost-master/groupbuy-bot/infrastructure/websocket/requirements.txt
aiohttp>=3.9
redis>=5.0
PyJWT>=2.8
